Revert "feat: Refactor monolithic script to modular governor package"
This reverts commit d82fe5b3a2685559ae2309c3b8a13e50d0d457bc.
