--- src/governor/constants.py
@@ -1,32 +1,0 @@
-from typing import Dict, Any
-
-class TemporalGoverningConstant:
-    """
-    Codifies the Rule of Intellectual Accounting: Design is Non-Temporal/Simultaneous;
-    Execution is Time-Bound/Sequential. This constant MUST NOT be confused with reality.
-    """
-    def __init__(self, rule_version: str = "1.0"):
-        self.rule_version: str = rule_version
-        self._is_sealed: bool = True
-
-        # The Rule: Intellectual Accounting (Non-Temporal Domain - אצילות)
-        self.INTELLECTUAL_ACCOUNTING: Dict[str, Any] = {
-            "DOMAIN": "World of Thought (אצילות)",
-            "TEMPORAL_STATE": "NON-TEMPORAL (Above Time)",
-            "RULE_LOGIC": "Outcome Shape determines Design; All stages eternally present ('One-Shot' Concept)."
-        }
-
-        # The Constraint: Sequential Execution (Temporal Domain - Reality)
-        self.REALITY_CONSTRAINT: Dict[str, Any] = {
-            "DOMAIN": "Sequential Execution (Reality)",
-            "TEMPORAL_STATE": "TIME-BOUND (Sequential, State-by-State)",
-            "EXECUTION_LOGIC": "Must minimize Quality Delta (ΔQ) as it builds over time."
-        }
-
-    def clarify_rule(self):
-        """Prints the clarification to prevent 'silly confusion'."""
-        print("\n--- GOVERNING CONSTANT CLARIFICATION (Preventing Confusion) ---")
-        print(f"RULE: The **{self.INTELLECTUAL_ACCOUNTING['DOMAIN']}** is {self.INTELLECTUAL_ACCOUNTING['TEMPORAL_STATE']}.")
-        print(f"CONSTRAINT: The **{self.REALITY_CONSTRAINT['DOMAIN']}** is {self.REALITY_CONSTRAINT['TEMPORAL_STATE']}.")
-        print("Conclusion: The Timeless Accounting MUST NOT Be Mistaken For The Time-Bound Execution.")
-        print("-------------------------------------------------------------")
