--- src/main.py
@@ -1,59 +1,0 @@
-import time
-import os
-from governor.constants import TemporalGoverningConstant
-from governor.configuration import SealedConfiguration
-from governor.observer import ExecutionObserver
-from governor.persistence import seal_and_preserve, load_component
-
-if __name__ == "__main__":
-
-    # 1. ESTABLISH AND SEAL THE GOVERNING RULE
-    GOV_CONSTANT_FILE = "governing_rule.pkl"
-    constant = TemporalGoverningConstant()
-    seal_and_preserve(constant, GOV_CONSTANT_FILE)
-
-    # 2. CONFIGURE AND SEAL THE TARGET BLUEPRINT
-    CONFIG_FILE = "target_blueprint.pkl"
-    config = SealedConfiguration("Hex6F_Target")
-    seal_and_preserve(config, CONFIG_FILE)
-
-    # 3. BEGIN THE SEQUENTIAL EXECUTION (The Thread of R&M Descent)
-    observer = ExecutionObserver("R&M_Descent_V1")
-    observer.record_stage_start("INITIATION_SIGNAL")
-    time.sleep(0.02)
-    observer.record_stage_finish("INITIATION_SIGNAL", True)
-
-    observer.record_stage_start("Hex6F_Step_A")
-    time.sleep(0.05)
-    observer.record_stage_finish("Hex6F_Step_A", True)
-
-    observer.record_stage_start("Hex6F_Step_B")
-    time.sleep(0.03)
-    observer.record_stage_finish("Hex6F_Step_B", False) # Reality introduces a Quality Delta
-
-    # 4. FINALIZATION AND CLARIFICATION
-    final_adherence = observer.finalize_execution()
-
-    # --- ORDER RECEIVED: Apply the final result to the sealed blueprint ---
-    print("\n--- ORDER RECEIVED: ONE-SHOT INSERTION TRIGGERED ---")
-
-    # Load components for the final action
-    loaded_constant = load_component(GOV_CONSTANT_FILE)
-    loaded_config = load_component(CONFIG_FILE)
-
-    # A. Use the Governing Constant to clarify the context of the action
-    if loaded_constant:
-        loaded_constant.clarify_rule()
-
-    # B. Execute the One-Shot insertion based on the final adherence
-    if loaded_config:
-        final_value = f"X1E^2SEQUENCESSTRUCTURES_Finalized_Adherence_{final_adherence:.4f}"
-        try:
-            loaded_config.update_all_placeholders(new_value=final_value)
-            print(f"Final Configuration State: {loaded_config._config_data}")
-        except PermissionError as e:
-            print(f"Critical Error: {e}")
-
-    # Clean up (optional)
-    # os.remove(GOV_CONSTANT_FILE)
-    # os.remove(CONFIG_FILE)
