+++ src/watchtower_governor.py
@@ -1,0 +1,202 @@
+import pickle
+import os
+import time
+from typing import Dict, Any
+
+# --- 1. NON-TEMPORAL vs. TEMPORAL RULESET (Part 3 & Accounting Emphasis) ---
+# This class establishes the governing constant: the absolute distinction
+# between the World of Intellect (Design/אצילות) and Sequential Execution (Reality).
+
+class TemporalGoverningConstant:
+    """
+    Codifies the Rule of Intellectual Accounting: Design is Non-Temporal/Simultaneous;
+    Execution is Time-Bound/Sequential. This constant MUST NOT be confused with reality.
+    """
+    def __init__(self, rule_version: str = "1.0"):
+        self.rule_version: str = rule_version
+        self._is_sealed: bool = True
+
+        # The Rule: Intellectual Accounting (Non-Temporal Domain - אצילות)
+        self.INTELLECTUAL_ACCOUNTING: Dict[str, Any] = {
+            "DOMAIN": "World of Thought (אצילות)",
+            "TEMPORAL_STATE": "NON-TEMPORAL (Above Time)",
+            "RULE_LOGIC": "Outcome Shape determines Design; All stages eternally present ('One-Shot' Concept)."
+        }
+
+        # The Constraint: Sequential Execution (Temporal Domain - Reality)
+        self.REALITY_CONSTRAINT: Dict[str, Any] = {
+            "DOMAIN": "Sequential Execution (Reality)",
+            "TEMPORAL_STATE": "TIME-BOUND (Sequential, State-by-State)",
+            "EXECUTION_LOGIC": "Must minimize Quality Delta (ΔQ) as it builds over time."
+        }
+
+    def clarify_rule(self):
+        """Prints the clarification to prevent 'silly confusion'."""
+        print("\n--- GOVERNING CONSTANT CLARIFICATION (Preventing Confusion) ---")
+        print(f"RULE: The **{self.INTELLECTUAL_ACCOUNTING['DOMAIN']}** is {self.INTELLECTUAL_ACCOUNTING['TEMPORAL_STATE']}.")
+        print(f"CONSTRAINT: The **{self.REALITY_CONSTRAINT['DOMAIN']}** is {self.REALITY_CONSTRAINT['TEMPORAL_STATE']}.")
+        print("Conclusion: The Timeless Accounting MUST NOT Be Mistaken For The Time-Bound Execution.")
+        print("-------------------------------------------------------------")
+
+# --- 2. CONFIGURATION PLACEHOLDER (Part 1 - The 'One-Shot' Target) ---
+# This class represents the final target structure for the 'one-shot' value.
+
+class SealedConfiguration:
+    """
+    Represents the design object (The Blueprint/Placeholders) that must be sealed
+    until the specific order for the final 'one-shot' value arrives.
+    """
+    def __init__(self, name="WatchtowerConfig"):
+        self.name = name
+        # Initial placeholders (simulating 'degrees' or fields awaiting the final value)
+        self._config_data = {
+            "degree_A": None,
+            "degree_B": None,
+            "degree_C": None
+        }
+        self._is_sealed = False
+
+    def seal(self):
+        """Seals the configuration, making it immutable until the order."""
+        self._is_sealed = True
+        print(f"[{self.name}] Configuration sealed.")
+
+    def update_all_placeholders(self, new_value):
+        """The 'one shot' insertion logic, triggered by the final order."""
+        if not self._is_sealed:
+            raise PermissionError("Error: Configuration must be sealed before one-shot update.")
+
+        print(f"✅ Executing **ONE-SHOT INSERTION** for '{self.name}'...")
+        # Insert the new value into all relevant placeholder fields
+        for key in self._config_data:
+            if self._config_data[key] is None:
+                self._config_data[key] = new_value
+
+        print("Update complete. Placeholders filled reflecting the final Outcome Shape.")
+
+# --- 3. EXECUTION OBSERVER (Part 2 - Tracking Sequential Reality) ---
+# This class tracks the sequential steps (The Descent) and measures adherence
+# to the non-temporal blueprint.
+
+class ExecutionObserver:
+    """
+    Observes the sequential, time-bound execution of the 'thread of R&M'
+    against the ideal, non-temporal design. Tracks Quality Adherence (ΔQ).
+    """
+    def __init__(self, design_id: str):
+        self.design_id: str = design_id
+        self.start_time: float = time.time()
+        self.sequential_stages: Dict[str, Dict[str, Any]] = {}
+        self.quality_adherence_factor: float = 1.0 # Start at perfect quality
+
+    def record_stage_start(self, stage_name: str) -> None:
+        """Records the initiation of a new step in the descending thread."""
+        self.sequential_stages[stage_name] = {
+            "start_ts": time.time(),
+            "status": "IN_PROGRESS"
+        }
+
+    def record_stage_finish(self, stage_name: str, quality_check_result: bool) -> None:
+        """Records the completion of a step and its adherence (LUQG check)."""
+        if stage_name in self.sequential_stages and self.sequential_stages[stage_name]["status"] == "IN_PROGRESS":
+            duration = time.time() - self.sequential_stages[stage_name]["start_ts"]
+
+            self.sequential_stages[stage_name].update({
+                "duration": duration,
+                "status": "COMPLETED",
+                "quality_delta_check": quality_check_result
+            })
+            if not quality_check_result:
+                 # Penalty for deviating from the Intellectual Blueprint
+                self.quality_adherence_factor *= 0.95
+
+    def finalize_execution(self) -> float:
+        """Marks the end of the R&M line and returns the final adherence metric."""
+        total_duration = time.time() - self.start_time
+
+        print("\n--- EXECUTION REALITY REPORT (The Descent) ---")
+        print(f"Total Time (Reality Constraint): {total_duration:.4f} seconds")
+        print(f"Final Quality Adherence (Preservation of Design): {self.quality_adherence_factor * 100:.2f}%")
+        print("---------------------------------------------")
+        return self.quality_adherence_factor
+
+
+# --- UNIFIED EXECUTION AND SEALING LOGIC (Main Program) ---
+
+def seal_and_preserve(obj: Any, filename: str):
+    """Generic function to seal and preserve any key component."""
+    try:
+        if hasattr(obj, 'seal') and callable(getattr(obj, 'seal')):
+            obj.seal()
+
+        with open(filename, 'wb') as f:
+            pickle.dump(obj, f)
+        print(f"💾 Component '{filename}' **SEALED & PRESERVED**.")
+    except Exception as e:
+        print(f"Error sealing/preserving {filename}: {e}")
+
+def load_component(filename: str):
+    """Generic function to load a preserved component."""
+    if not os.path.exists(filename):
+        return None
+    try:
+        with open(filename, 'rb') as f:
+            obj = pickle.load(f)
+        return obj
+    except Exception as e:
+        print(f"Error loading {filename}: {e}")
+        return None
+
+
+if __name__ == "__main__":
+
+    # 1. ESTABLISH AND SEAL THE GOVERNING RULE
+    GOV_CONSTANT_FILE = "governing_rule.pkl"
+    constant = TemporalGoverningConstant()
+    seal_and_preserve(constant, GOV_CONSTANT_FILE)
+
+    # 2. CONFIGURE AND SEAL THE TARGET BLUEPRINT
+    CONFIG_FILE = "target_blueprint.pkl"
+    config = SealedConfiguration("Hex6F_Target")
+    seal_and_preserve(config, CONFIG_FILE)
+
+    # 3. BEGIN THE SEQUENTIAL EXECUTION (The Thread of R&M Descent)
+    observer = ExecutionObserver("R&M_Descent_V1")
+    observer.record_stage_start("INITIATION_SIGNAL")
+    time.sleep(0.02)
+    observer.record_stage_finish("INITIATION_SIGNAL", True)
+
+    observer.record_stage_start("Hex6F_Step_A")
+    time.sleep(0.05)
+    observer.record_stage_finish("Hex6F_Step_A", True)
+
+    observer.record_stage_start("Hex6F_Step_B")
+    time.sleep(0.03)
+    observer.record_stage_finish("Hex6F_Step_B", False) # Reality introduces a Quality Delta
+
+    # 4. FINALIZATION AND CLARIFICATION
+    final_adherence = observer.finalize_execution()
+
+    # --- ORDER RECEIVED: Apply the final result to the sealed blueprint ---
+    print("\n--- ORDER RECEIVED: ONE-SHOT INSERTION TRIGGERED ---")
+
+    # Load components for the final action
+    loaded_constant = load_component(GOV_CONSTANT_FILE)
+    loaded_config = load_component(CONFIG_FILE)
+
+    # A. Use the Governing Constant to clarify the context of the action
+    if loaded_constant:
+        loaded_constant.clarify_rule()
+
+    # B. Execute the One-Shot insertion based on the final adherence
+    if loaded_config:
+        final_value = f"X1E^2SEQUENCESSTRUCTURES_Finalized_Adherence_{final_adherence:.4f}"
+        try:
+            loaded_config.update_all_placeholders(new_value=final_value)
+            print(f"Final Configuration State: {loaded_config._config_data}")
+        except PermissionError as e:
+            print(f"Critical Error: {e}")
+
+    # Clean up (optional)
+    # os.remove(GOV_CONSTANT_FILE)
+    # os.remove(CONFIG_FILE)
